Working

+++ include/constats.h
@@ -15,18 +15,18 @@ typedef struct stats_t
 {
 	uint64_t N;			// The size of the sample set
 
-	float mean;			// The mean of the data
-	float stdev;		// The standard deviation
-	float abdev;		// The mean absolute deviation
+	double mean;		// The mean of the data
+	double stdev;		// The standard deviation
+	double abdev;		// The mean absolute deviation
 	int64_t min;		// The minimum value
 	int64_t max;		// The maximum value
 
 	int64_t tolerance;	// The tolerance level used in classifying outliers
 	uint64_t outliers;	// The number of data points classified as outliers
 
-	float norm_mean;	// The mean of the data without the outliers
-	float norm_stdev;	// The standard deviation without the outliers
-	float norm_abdev;	// The mean absolute deviation without the outliers
+	double norm_mean;	// The mean of the data without the outliers
+	double norm_stdev;	// The standard deviation without the outliers
+	double norm_abdev;	// The mean absolute deviation without the outliers
 	int64_t norm_min;	// The minimum value without the outliers
 	int64_t norm_max;	// The maximum value without the outliers
